Move xml-parsing methods to libvirt-xml-parse.js
