+++ src/libvirt-xml-parse.js
