+++ src/libvirt-xml-update.js
@@ -1,4 +1,4 @@
-import { getDoc, getSingleOptionalElem } from './libvirt-common.js';
+import { getDoc, getSingleOptionalElem } from './libvirt-xml-parse.js';
 import { getNextAvailableTarget, logDebug } from './helpers.js';
 
 export function updateDisk({ domXml, diskTarget, readonly, shareable, busType, existingTargets, cache }) {
